+++ src/pages/OffersPage/components/OffersTabContent/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './useLenderLoansAndOffers'
+export * from './useOffersTabContent'
